Refactor: Move Snack model to models directory and remove unused CartItem and Product files
