+++ lib/domain/state/app_notifier.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:snackautomat_2502/domain/snack.dart';
+import 'package:snackautomat_2502/models/snack.dart';
 import 'package:snackautomat_2502/domain/state/app_state.dart';
 import 'package:snackautomat_2502/models/coinstack.dart';
