+++ lib/models/snack.dart
